# OneDrive/Imagens/Dot-Explorer/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Santa Print — Sistema de Controle Operacional Interno
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardResumo {
  total: number;
  emAnalise: number;
  concluidas: number;
  canceladas: number;
  aguardandoPeca: number;
  paraConserto: number;
  hojeAbertas: number;
  hojeFechadas: number;
}

export interface StatusCount {
  situacao: string;
  total: number;
}

export interface EquipamentoCount {
  tipo: string;
  total: number;
}

export interface Chamada {
  empresa_number?: number;
  id_sales?: string;
  cod_situacao_zenthi?: number;
  status?: number;
  situacao?: string;
  situacao_zenthi: string;
  codigo: number;
  /** @nullable */
  valor_chamada?: number | null;
  emissao: string;
  /** @nullable */
  encerramento?: string | null;
  /** @nullable */
  defeito_informado?: string | null;
  /** @nullable */
  defeito_verificado?: string | null;
  /** @nullable */
  cancelamento?: string | null;
  /** @nullable */
  motivo_cancelamento?: string | null;
  /** @nullable */
  data_inicio_servico?: string | null;
  /** @nullable */
  data_final_servico?: string | null;
  /** @nullable */
  observacao?: string | null;
  /** @nullable */
  servico_realizado?: string | null;
  /** @nullable */
  observacao_servico?: string | null;
  cliente_codigo?: number;
  razao_social: string;
  /** @nullable */
  obs_cliente?: string | null;
  /** @nullable */
  cpf_cnpj?: string | null;
  cidade: string;
  /** @nullable */
  docto_tecnico?: string | null;
  /** @nullable */
  email_tecnico?: string | null;
  /** @nullable */
  codigo_ocorrencia?: string | null;
  /** @nullable */
  descricao_ocorrencia?: string | null;
  /** @nullable */
  contrato?: number | null;
  /** @nullable */
  codigo_tipo_contrato?: string | null;
  /** @nullable */
  tipo_contrato?: string | null;
  /** @nullable */
  codigo_tipo_entrada?: string | null;
  /** @nullable */
  tipo_entrada?: string | null;
  /** @nullable */
  produto_os?: string | null;
  /** @nullable */
  marca?: string | null;
  /** @nullable */
  modelo?: string | null;
  /** @nullable */
  numero_serie?: string | null;
  /** @nullable */
  setor?: string | null;
  /** @nullable */
  suporte?: string | null;
  /** @nullable */
  ibge?: number | null;
  /** @nullable */
  data_prevista?: string | null;
  /** @nullable */
  tipo_equipamento?: number | null;
  /** @nullable */
  desc_tipo_equipamento?: string | null;
  /** @nullable */
  marca_codigo?: number | null;
  /** @nullable */
  setor_codigo?: number | null;
  /** @nullable */
  ativo?: string | null;
}

export interface ChamadasPaginadas {
  data: Chamada[];
  total: number;
  page: number;
  limit: number;
}

export interface Peca {
  empresa_number?: number;
  filial_number?: number;
  chamada_number: number;
  /** @nullable */
  data_abertura?: string | null;
  /** @nullable */
  cliente_number?: number | null;
  /** @nullable */
  situacao?: number | null;
  /** @nullable */
  data_encerramento?: string | null;
  /** @nullable */
  data_prevista?: string | null;
  /** @nullable */
  entrada?: string | null;
  /** @nullable */
  garantia?: string | null;
  /** @nullable */
  numero_os?: string | null;
  /** @nullable */
  nota_fiscal?: number | null;
  /** @nullable */
  serie?: string | null;
  /** @nullable */
  produto_principal?: string | null;
  /** @nullable */
  descricao_equip?: string | null;
  /** @nullable */
  modelo_equip?: string | null;
  /** @nullable */
  marca_number?: number | null;
  /** @nullable */
  desc_marca?: string | null;
  /** @nullable */
  defeito_equip?: string | null;
  /** @nullable */
  defeito_verificado?: string | null;
  /** @nullable */
  servicos_realizar?: string | null;
  /** @nullable */
  obs_servicos?: string | null;
  /** @nullable */
  login_number?: number | null;
  /** @nullable */
  id_sales?: string | null;
  /** @nullable */
  produto?: string | null;
  desc_produto: string;
  /** @nullable */
  qtdem?: number | null;
  /** @nullable */
  preco_unitario?: number | null;
  /** @nullable */
  valor_item?: number | null;
  /** @nullable */
  valor_custo?: number | null;
  /** @nullable */
  movimentou_estoque?: string | null;
  id_sales_peca: string;
}

export interface PecasPaginadas {
  data: Peca[];
  total: number;
  page: number;
  limit: number;
}

export interface Pedido {
  /** @nullable */
  empresa?: number | null;
  /** @nullable */
  pedido?: string | null;
  /** @nullable */
  nome_repres?: string | null;
  /** @nullable */
  mes_ano?: string | null;
  /** @nullable */
  cliente?: number | null;
  /** @nullable */
  nome_cliente?: string | null;
  /** @nullable */
  pedido_descricao?: string | null;
  /** @nullable */
  produto?: string | null;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  qtde?: number | null;
  /** @nullable */
  valor_bruto?: number | null;
  /** @nullable */
  valor_item?: number | null;
  /** @nullable */
  valor_desconto?: number | null;
  /** @nullable */
  valor_venda?: number | null;
  /** @nullable */
  valor_custo?: number | null;
  /** @nullable */
  margem?: number | null;
  /** @nullable */
  data_emissao?: string | null;
  /** @nullable */
  forma_pagto?: string | null;
  /** @nullable */
  prazo?: string | null;
}

export interface PedidosPaginados {
  data: Pedido[];
  total: number;
  page: number;
  limit: number;
}

export interface Contrato {
  /** @nullable */
  id?: number | null;
}

export interface ContratosPaginados {
  data: Contrato[];
  total: number;
  page: number;
  limit: number;
}

export interface FiltrosOpcoes {
  situacoes: string[];
  tiposEquipamento: string[];
  marcas: string[];
  cidades: string[];
  tecnicos: string[];
  tiposContrato: string[];
  tiposEntrada: string[];
}

export type GetDashboardResumoParams = {
dataInicio?: string;
dataFim?: string;
};

export type GetDashboardPorStatusParams = {
dataInicio?: string;
dataFim?: string;
};

export type GetDashboardPorEquipamentoParams = {
dataInicio?: string;
dataFim?: string;
};

export type ListChamadasParams = {
/**
 * Status: Em Análise, Concluído, Cancelado, Aguardando Peça, Para Conserto
 */
situacao?: string;
dataInicio?: string;
dataFim?: string;
/**
 * MULTIFUNCIONAL, IMPRESSORA, SCANNERS, ETIQUETA, ZEBRA
 */
tipoEquipamento?: string;
/**
 * Contrato, Normal
 */
tipoContrato?: string;
/**
 * Interna, Externa
 */
tipoEntrada?: string;
marca?: string;
cidade?: string;
/**
 * Busca por razao_social
 */
cliente?: string;
/**
 * Busca por email_tecnico
 */
tecnico?: string;
page?: number;
limit?: number;
};

export type ListPecasParams = {
chamadaNumero?: number;
descProduto?: string;
marca?: string;
dataInicio?: string;
dataFim?: string;
page?: number;
limit?: number;
};

export type ListPedidosParams = {
/**
 * Formato MM/YYYY
 */
mesAno?: string;
cliente?: string;
representante?: string;
formaPagto?: string;
page?: number;
limit?: number;
};

export type ListContratosParams = {
tipoContrato?: string;
page?: number;
limit?: number;
};