# OneDrive/Imagens/Dot-Explorer/lib/api-zod/src/generated/types/chamada.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Santa Print — Sistema de Controle Operacional Interno
 * OpenAPI spec version: 0.1.0
 */

export interface Chamada {
  empresa_number?: number;
  id_sales?: string;
  cod_situacao_zenthi?: number;
  status?: number;
  situacao?: string;
  situacao_zenthi: string;
  codigo: number;
  /** @nullable */
  valor_chamada?: number | null;
  emissao: string;
  /** @nullable */
  encerramento?: string | null;
  /** @nullable */
  defeito_informado?: string | null;
  /** @nullable */
  defeito_verificado?: string | null;
  /** @nullable */
  cancelamento?: string | null;
  /** @nullable */
  motivo_cancelamento?: string | null;
  /** @nullable */
  data_inicio_servico?: string | null;
  /** @nullable */
  data_final_servico?: string | null;
  /** @nullable */
  observacao?: string | null;
  /** @nullable */
  servico_realizado?: string | null;
  /** @nullable */
  observacao_servico?: string | null;
  cliente_codigo?: number;
  razao_social: string;
  /** @nullable */
  obs_cliente?: string | null;
  /** @nullable */
  cpf_cnpj?: string | null;
  cidade: string;
  /** @nullable */
  docto_tecnico?: string | null;
  /** @nullable */
  email_tecnico?: string | null;
  /** @nullable */
  codigo_ocorrencia?: string | null;
  /** @nullable */
  descricao_ocorrencia?: string | null;
  /** @nullable */
  contrato?: number | null;
  /** @nullable */
  codigo_tipo_contrato?: string | null;
  /** @nullable */
  tipo_contrato?: string | null;
  /** @nullable */
  codigo_tipo_entrada?: string | null;
  /** @nullable */
  tipo_entrada?: string | null;
  /** @nullable */
  produto_os?: string | null;
  /** @nullable */
  marca?: string | null;
  /** @nullable */
  modelo?: string | null;
  /** @nullable */
  numero_serie?: string | null;
  /** @nullable */
  setor?: string | null;
  /** @nullable */
  suporte?: string | null;
  /** @nullable */
  ibge?: number | null;
  /** @nullable */
  data_prevista?: string | null;
  /** @nullable */
  tipo_equipamento?: number | null;
  /** @nullable */
  desc_tipo_equipamento?: string | null;
  /** @nullable */
  marca_codigo?: number | null;
  /** @nullable */
  setor_codigo?: number | null;
  /** @nullable */
  ativo?: string | null;
}