# OneDrive/Imagens/Dot-Explorer/lib/api-zod/src/generated/types/contratosPaginados.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Santa Print — Sistema de Controle Operacional Interno
 * OpenAPI spec version: 0.1.0
 */
import type { Contrato } from './contrato';

export interface ContratosPaginados {
  data: Contrato[];
  total: number;
  page: number;
  limit: number;
}