# OneDrive/Imagens/Dot-Explorer/lib/api-zod/src/generated/types/pecasPaginadas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Santa Print — Sistema de Controle Operacional Interno
 * OpenAPI spec version: 0.1.0
 */
import type { Peca } from './peca';

export interface PecasPaginadas {
  data: Peca[];
  total: number;
  page: number;
  limit: number;
}